+++ src/Utility/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add exception code `1655559296` to `\Heptacom\HeptaConnect\Utility\ClassString\Contract\SubtypeClassStringContract::__construct` when the given class string is not of the expected type
 - Add `\Heptacom\HeptaConnect\Utility\ClassString\UnsafeClassString` based on class `\Heptacom\HeptaConnect\Utility\ClassString\Contract\ClassStringReferenceContract` with `\Heptacom\HeptaConnect\Utility\ClassString\ClassStringReferenceCollection` and `\Heptacom\HeptaConnect\Utility\ClassString\AbstractClassStringReferenceCollection` to have a string references, that could be a class-string for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
 - Add `\Heptacom\HeptaConnect\Utility\ClassString\Exception\InvalidClassNameException`, `\Heptacom\HeptaConnect\Utility\ClassString\Exception\InvalidSubtypeClassNameException` and `\Heptacom\HeptaConnect\Utility\ClassString\Exception\UnexpectedLeadingNamespaceSeparatorInClassNameException` to reference class-string issues for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
+- Add `\Heptacom\HeptaConnect\Utility\Php\ClassCodeHasher` to calculate unique hashes for code of classes, interfaces and traits
 
 ### Changed